backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyJWT==2.9.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.8.2
gunicorn==23.0.0
